Remove duplicate file and run from build_artefact
